dsada

+++ nanos-lite/src/device.c
@@ -15,8 +15,9 @@ static const char *keyname[256] __attribute__((used)) = {
 };
 
 size_t serial_write(const void *buf, size_t offset, size_t len) {
-  for (size_t i = 0; i < len; i++)
+  for (int i = 0; i < len; i++)
   {
+    //putch(*((char *)buf[i]));
     putch(*((char *)buf + i));
   }
   return len;
